apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
redis
anthropic
openai
langchain
langchain-text-splitters
pypdf